requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
alembic
psycopg2-binary
python-jose[cryptography]
python-multipart
pydantic[email]
pytest
httpx
passlib[bcrypt]==1.7.4
bcrypt==4.3.0